✅ import new classes and set fps
